# doc/swagger.json
{"openapi":"3.0.0","paths":{"/user":{"get":{"operationId":"UsersController_getAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}}},"/user/{id}":{"get":{"operationId":"UsersController_getOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"UsersController_updatePassword","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/track":{"get":{"operationId":"TracksController_getAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"TracksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTrackDto"}}}},"responses":{"201":{"description":""}}}},"/track/{id}":{"get":{"operationId":"TracksController_getOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"TracksController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTrackDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"TracksController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/artist":{"get":{"operationId":"ArtistsController_getAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"ArtistsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateArtistDto"}}}},"responses":{"201":{"description":""}}}},"/artist/{id}":{"get":{"operationId":"ArtistsController_getOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"ArtistsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateArtistDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ArtistsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/album":{"get":{"operationId":"AlbumsController_getAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"AlbumsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlbumDto"}}}},"responses":{"201":{"description":""}}}},"/album/{id}":{"get":{"operationId":"AlbumsController_getOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"AlbumsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAlbumDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AlbumsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/favs":{"get":{"operationId":"FavoritesController_getAll","parameters":[],"responses":{"200":{"description":""}}}},"/favs/track/{id}":{"post":{"operationId":"FavoritesController_addTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavoritesController_deleteTrack","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/favs/album/{id}":{"post":{"operationId":"FavoritesController_addAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavoritesController_deleteAlbum","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/favs/artist/{id}":{"post":{"operationId":"FavoritesController_addArtis","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}},"delete":{"operationId":"FavoritesController_deleteArtist","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Service NodeJS","description":"The service API description","version":"1.0.0","contact":{}},"tags":[{"name":"service","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdatePasswordDto":{"type":"object","properties":{}},"CreateTrackDto":{"type":"object","properties":{}},"UpdateTrackDto":{"type":"object","properties":{}},"CreateArtistDto":{"type":"object","properties":{}},"UpdateArtistDto":{"type":"object","properties":{}},"CreateAlbumDto":{"type":"object","properties":{}},"UpdateAlbumDto":{"type":"object","properties":{}}}}}